requirements.txt
langchain
langchain-anthropic
wikipedia
langchain-community
langchain-openai
pydantic
python-dotenv
duckduckgo-search
google-search-results
fastapi
uvicorn
